ecommerce_db/requirements.txt
Django==5.2
djangorestframework
dj-database-url
python-dotenv
psycopg2-binary
whitenoise
gunicorn
scikit-surprise
scikit-learn
numpy
pandas
requests